feat: update login structure and token management
