+++ providers/auth-provider/login.ts
@@ -1,6 +1,8 @@
 import axios from 'axios';
 
-export type Strategy = 'spa' | 'native';
+import { getAxios } from './utils';
+
+import { Strategy } from '~/types/strategy';
 
 export type LoginParams = {
   email: string;
@@ -9,36 +11,15 @@ export type LoginParams = {
   strategy: Strategy;
 };
 
-export const login = async (params: LoginParams): Promise<{ token: string }> => {
-  const baseUrl = process.env.EXPO_PUBLIC_API_URL;
-
-  let csrfCookieRoute = undefined;
-  let headers = {};
-  let route = '';
-  if (params.strategy === 'spa') {
-    csrfCookieRoute = 'sanctum/csrf-cookie';
-    headers = {
-      'X-Requested-With': 'XMLHttpRequest',
-    };
-    route = 'login';
-  } else {
-    headers = {
-      'Content-Type': 'application/json',
-    };
-    route = 'api/v1/login';
-  }
-
-  const http = axios.create({
-    baseURL: baseUrl,
-    headers,
-    withCredentials: true,
-    withXSRFToken: true,
-  });
-
+export const login = async (params: LoginParams): Promise<{ token: string } | void> => {
   try {
-    if (csrfCookieRoute) {
-      await http.get(csrfCookieRoute);
+    const http = getAxios(params.strategy);
+
+    if (params.strategy === 'spa') {
+      await http.get('sanctum/csrf-cookie');
     }
+
+    const route = params.strategy === 'spa' ? 'login' : 'api/v1/login';
 
     const result = await http.post(route, {
       email: params.email,
@@ -46,9 +27,7 @@ export const login = async (params: LoginParams): Promise<{ token: string }> => 
       device_name: params.deviceName,
     });
 
-    if (params.strategy === 'spa') {
-      return { token: 'token' };
-    } else {
+    if (params.strategy === 'native') {
       return { token: result.data['token'] };
     }
   } catch (e) {
