+++ types/strategy.d.ts
@@ -1,0 +1,1 @@
+export type Strategy = 'spa' | 'native';
